Slight refactor and gdt and idt stubs

+++ include/string.h
@@ -7,8 +7,8 @@
 int strcmp(const char *, const char *);
 size_t strlen(const char *);
 const char *strchr(const char *, int);
-void *memcpy(void *, const void *, uint32_t);
-void *memmove(void *, const void *, uint32_t);
-void *memset(void *, int, uint32_t);
+void *memcpy(void *, const void *, size_t);
+void *memmove(void *, const void *, size_t);
+void *memset(void *, int, size_t);
 
 #endif // BEAVER_STRING_H
